+++ src/main.js
@@ -1,7 +1,8 @@
 import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
+import App from "@/App.vue";
+import router from "@/router";
 
+import "@/plugins/lazyload";
 Vue.config.productionTip = false;
 
 new Vue({