+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## Version 0.0.3 (2020-07-29)
+
+Do performance optimization
+- gzip: compress large files of js, css, html, and json
+- picture: 
+  - compress picture files of jpg, jpeg, png, gif, and svg
+  - add lazy loader
+- json: minify static json files
+- analyzer: add webpack-bundle-analyzer
+- dns: prefetch dns
+
 ## Version 0.0.2 (2020-07-24)
 
 Do configurations